Added the custom authentication
