+++ account/urls.py
@@ -3,13 +3,13 @@ from django.urls import path
 # Views
 from .views import (
     CreateUserAccountView,
-    GetUserAccountView,
+    RetrieveUserAccountView
 )
 
-from .authentication import TokenAuthentication
+from .auth.views import AuthenticateUserAccountView
 
 urlpatterns = [
-    path('', GetUserAccountView.as_view(), name="account:get")
-    path('create/', CreateUserAccountView.as_view(), name="account:create")
-    path('signin/', TokenAuthentication.as_view(), name="account:signin")
+    path('', RetrieveUserAccountView.as_view(), name="account_get"),
+    path('create/', CreateUserAccountView.as_view(), name="account_create"),
+    path('signin/', AuthenticateUserAccountView.as_view(), name="account_signin")
 ]
